basicagent/pyproject.toml
[project]
name = "hello-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agentops>=0.4.12",
    "openai-agents>=0.0.4",
    "python-dotenv>=1.0.1",
]
